requirements.txt
aioserial
pyserial
inquirerpy